# [IMP] core: simplify implementation of SQL wrapper