Simply don't keep the `SQL` tree, and merge the arguments directly into the code and parameters.  This makes the `SQL` wrapper simpler, faster, and it does not suffer for recursion limit errors.
```py
from odoo.tools import SQL

%timeit (sql := SQL("hello world %s, %s", 4, "ok")), sql.code, sql.params
# before: 3.85 µs ± 11.6 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
# after : 1.35 µs ± 1.14 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

%timeit (sql := SQL("select * FROM %s WHERE %s", SQL.identifier("ok"), SQL('x > %s', 5))), sql.code, sql.params
# before: 7.81 µs ± 161 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
# after : 3.14 µs ± 45.7 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

%timeit (sql := SQL(",").join([SQL('ok'), *range(30), SQL("qbc%s", 5)])), sql.code, sql.params
# before: 25 µs ± 108 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
# after : 9.97 µs ± 22 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

def complex_sql(size=20):
    if size > 0:
        return SQL("(%s, %s)", complex_sql(size - 1), complex_sql(size - 2))
    return SQL("res")

%timeit (sql := complex_sql()), sql.code, sql.params
# before: 68.5 ms ± 228 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
# after : 29 ms ± 33.6 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```