@robodoo 2024-07-19
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/173794.png)](https://mergebot.odoo.com/odoo/odoo/pull/173794)