@robodoo 2024-07-20
@rco-odoo staging failed: ci/l10n on f94cca78e9e7f03a4443555712f8f77197161e66 (view more at https://runbot.odoo.com/runbot/build/65788663)