@robodoo 2024-07-20
@rco-odoo 'ci/runbot' failed on this reviewed PR.